app/requirements.txt
# Core
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.2
Flask-Cors==4.0.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9

# Email
emails==0.6
premailer==3.10.0

# Security
bcrypt==4.0.1
PyJWT==2.8.0

# Utilities
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3

# Development
debugpy==1.8.0
pytest==7.4.2
pytest-cov==4.1.0
black==23.9.1
flake8==6.1.0

# API
gunicorn==21.2.0
Werkzeug==2.3.7
Flask-Mail==0.9.1
Flask-Migrate==4.0.5

# Documentation
Flask-Swagger-UI==4.11.1
apispec==6.3.1
marshmallow==3.20.1

pandas
Flask-Login==0.6.3
